Improve Terminal.lines semantics
Terminal.lines used to return the number of lines on screen.

It now returns how many lines of text from the file to display that can
be displayed.

The old definition felt more like an internal number, and the new
definition feels more like something the user of the Terminal would like
to have.