Don't find lines matching escape codes only
Previously we fixed *highlighting* of escaped lines.

This change makes sure we don't say "There's a search hit on this line",
but without highlighting anything since the hit was part of the escape
code.

We've also refactored our AnsiUtils module into an actual class
containing the same functionality.

Moar now keeps an array of AnsiStrings rather than an array of strings.