Handle unknown versions of Moar better
The Moar version is maintained in the version control system (Git), and
if somebody gets Moar through some other means or just copies moar.rb
out of the source tree then we won't have any version information.

Sources should be retrieved using 'git clone' and installs should be
done using 'rake'.

This change makes Moar say that the version is 'UNKNOWN' if it's
unknown, and it will inform users about Git (for getting the sources)
and Rake (for installing).