Don't remember last search when hitting '/' again
This basically reverts an older change.

The reason is that after using this for real for quite some time now,
I concluded that remembering the last search feels weird most of the
time.

The Less maintains a search history, but my feeling (after using both
Less and Moar a lot) is that that's not really necessary when you have
interactive search.

Change-Id: Iba75b9eb1b43dbd82d9286a45bdd962556ff5501