Modify the highlighting API
This changes the highlighting API to take a string instead of a file
because strings are what Chroma can highlight.

With this change in place, we should invoke the highlighter after
we're done reading the input. This will allow highlighting both files
(as before) and streaming input.

This also moves the responsibility of checking the file against the max
highlight size from the highlighter to the caller.