Hide line numbers on man pages
By default. Press left arrow to get them back. Or use start moar with
--no-linenumbers=false.

`man` starts by checking the terminal width, and then formats man pages
to fit that width.

Then, if moar adds line numbers, the rightmost part of the man page
will be shifted out of view.

So we try to detect showing man pages, and in that case disable line
numbers so that the rightmost part of the page is visible by default.

Fixes #200.