Improve search hits highlighting
Before this change, if you had a search hit inside inverse video text
with a non standard color, search hits in there would be invisible.

With this change in place, we now force the search hits highlights to
use the default colors, making them visible in this case.

This could sometimes be a problem when searching riff highlighted git
diffs:
https://github.com/walles/riff