Show Fish / Powershell specific help
On `moar --help`, if `moar` is not the default pager, try identifying
the user's shell and print instructions for setting `PAGER` in that
shell.

Identifies fish or Powershell, then falls back on bash / zsh
instructions (since I failed to come up with a sensible way of
identifying those).

Before this change, we always printed bash / zsh instructions.

Fixes: https://github.com/walles/moar/issues/251