Let ESC skip back to where the search started
When searching in the pager, pressing ESC will now skip back to where
the search started. This is useful when you want to quickly jump back to
where you started searching from.

Pressing RETURN will still stop the search and leave the cursor where it
is.

Being able to press ESC to go back somewhat improves on the first half
of https://github.com/walles/moar/issues/257, related to searching from
the current position to the end.