Don't log unexpected exits
Before this change, if we crashed inside either of these main loops, we
would still log messages about exiting the loops. Even if the reason for
exiting was a panic().

With this change in place, we only log expected exits. The idea is that
this should make it easier to understand where the crash was.