Detect compressed files by magic numbers
Rather than by file extension. This makes us handle misnamed files
correctly.

Also fix a bug in how we display the filename in the UI. Before this
change, we just displayed the plain file name even if we uncompressed
it.

With this change in place, we now display the filename without any
compression extension.