Start fixing line number prefix length calculation
Next step would be to test having it giving every line the same number
and verify the line numbers get the right amount of space.

Try this and go to the end: moar /etc/services

Check that the line numbers are 99 all the way.

After that, we should work on filtering.