Improve locking
Before this change, if two threads at the same time called Plain(), both
would do the computation.

With this change in place, only one of them will do the computation,
while the other one is waiting. After the computation is done, the
second thread will just get the result from the first one.