Make the search benchmark much faster to run
1s rather than 55s.

Before this change, the execution time for each iteration was
proportional to the number of iterations, which made the whole thing
be quadratic in time complexity.

With this change in place, the execution time of each iteration is
linear in the number of iterations.

That enables Go to decide how many iterations to run for the whole thing
to take about 1s.

Thanks to GitHub Copilot for suggesting the fix!